pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strategy-engine"
version = "0.1.0"
description = "Open-source quantitative allocation signal engine for multi-asset portfolio management"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "strategy-builder-ai contributors"},
]
keywords = ["quantitative-finance", "trading", "signals", "portfolio", "dca", "z-score"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "pydantic>=2.5",
    "httpx>=0.25",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3",
    "hmmlearn>=0.3",
    "statsmodels>=0.14",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "celery[redis]>=5.3",
    "python-jose[cryptography]>=3.3",
    "siwe>=4.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
all = ["strategy-engine[ml,api,dev]"]

[project.urls]
Homepage = "https://github.com/azank1/strategy-builder-ai"
Repository = "https://github.com/azank1/strategy-builder-ai"

[tool.hatch.build.targets.wheel]
packages = ["strategy_engine"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=strategy_engine --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
